requirements.txt
streamlit
moviepy==2.1.2
python-ffmpeg==1.0.16
openai
python-dotenv
numpy
Pillow
requests
PyPDF2
python-docx
beautifulsoup4
pandas